refactor(Date detail): new DateCard component
